refactor(LineService): 상행역에서 하행역까지 모든 역을 가져온다
